+++ test/result-te/test_and.js
@@ -47,7 +47,7 @@ describe('Result<T, E>.and()', function(){
             });
 
             it('the returned value should be `Ok', function () {
-                assert.strictEqual(result.isOk(), true);
+                assert.strictEqual(result.isOk, true);
             });
 
             it('the returned value should wrap the expected value', function () {
@@ -70,7 +70,7 @@ describe('Result<T, E>.and()', function(){
             });
 
             it('the returned value should be `Err', function () {
-                assert.strictEqual(result.isErr(), true);
+                assert.strictEqual(result.isErr, true);
             });
 
             it('the returned value should wrap the expected value', function () {
@@ -95,7 +95,7 @@ describe('Result<T, E>.and()', function(){
             });
 
             it('the returned value should be `Err', function () {
-                assert.strictEqual(result.isErr(), true);
+                assert.strictEqual(result.isErr, true);
             });
 
             it('the returned value should wrap the expected value', function () {
@@ -118,7 +118,7 @@ describe('Result<T, E>.and()', function(){
             });
 
             it('the returned value should be `Err', function () {
-                assert.strictEqual(result.isErr(), true);
+                assert.strictEqual(result.isErr, true);
             });
 
             it('the returned value should wrap the expected value', function () {